+++ demo/src/main/java/im/status/keycard/app/MainActivity.java
@@ -8,7 +8,7 @@ import im.status.keycard.demo.R;
 import im.status.keycard.io.CardChannel;
 import im.status.keycard.io.CardListener;
 import im.status.keycard.io.CardManager;
-import im.status.keycard.wallet.*;
+import im.status.keycard.keycard.*;
 import org.spongycastle.util.encoders.Hex;
 
 public class MainActivity extends AppCompatActivity {
