+++ lib/src/main/java/im/status/keycard/keycard/BIP32KeyPair.java
@@ -1,4 +1,4 @@
-package im.status.keycard.wallet;
+package im.status.keycard.keycard;
 
 import org.spongycastle.crypto.digests.KeccakDigest;
 import org.spongycastle.math.ec.ECPoint;
