+++ lib/src/main/java/im/status/keycard/keycard/KeyPath.java
@@ -1,4 +1,4 @@
-package im.status.keycard.wallet;
+package im.status.keycard.keycard;
 
 import java.util.StringTokenizer;
