+++ lib/src/main/java/im/status/keycard/keycard/SecureChannelSession.java
@@ -1,4 +1,4 @@
-package im.status.keycard.wallet;
+package im.status.keycard.keycard;
 
 import im.status.keycard.io.APDUCommand;
 import im.status.keycard.io.APDUException;
